Deduplicate PDSLabel/PDS4Label test mock factories via conftest.py
test_classes_label.py and test_classes_label_pds4.py each had a
verbatim copy of make_context_products/make_bundle/make_collection/
make_product/make_setup_pds4 (both classes need identical setup/
product mocks). SonarCloud flagged this on PR #308 as 8.3% new-code
duplication, entirely attributable to test_classes_label_pds4.py's
85 duplicated lines against test_classes_label.py.

Moved the factories into a new label_test_helpers fixture in
conftest.py, following the existing base_helpers pattern already used
by the leaf label test files. Both files, and every fixture/test that
previously called the bare module-level functions directly, now
request label_test_helpers and call the equivalent method on it.

pytest tests/npb/ -> 1764 passed, 8 skipped, 1 xfailed, 99% coverage,
unchanged from before this commit.
